# Speed up quantification optimizations by unswitching