@milseman 2023-12-15
This is now superseded by a larger PR that includes much more cleanup, refactoring, and preparation for more improvements.